refactor operation data
